Select role from AWS_PROFILE env variable
When the AWS_PROFILE environment variable is set, use that to select
the role_arn, rather than prompting the user to select a role. This
addresses an issue where a user-selected role and the AWS_PROFILE
do not match, and credentials are then saved associated to the wrong
role name.