Move configuration settings to a Config object (#61)
This is a breaking change, since configuration elements in the ini file,
command-line, and environment variables have been renamed.

All the settings that were accessed in `settings.*` have been moved to
config object, automatically instatiated from the Config class.
Elements can be accessed by the section name, so that they easier to
track.

External documentation has not been updated, as there is parallel work
to update all documentation for the upcoming 2.0 release.

Co-authored-by: Yaroslav Kudrin <Iaroslav_Kudrin@epam.com>