Clean up ini parsing, early logging (#76)
This changeset introduces "early" logging, so that initialilzation can
be observed as well. We refactored ini file reading and writing so that
we can reuse the same code, as well as getting information from the user
interactively. A variety of functions now have useful return values, so
that they can be tested. There is also a few more unit tests to improve
our confidene on the codebase.

This introduces tests to ensure #64 is no longer an issue.