Centralize verification and sanitization methods (#81)
We discovered through testing that setting app_url and org_url to
different FQDNs would cause unexpected issue. This was relatively easy
for a user to trigger if they had a base profile set to a main Okta
Organization, and and a downstream profile to an Okta Preview Org.

To address that issue, this changeset normalizes input validation, pulls
it into a set of methods that are called sequentially, and performs
pre-flight checks before we attempt to call any remote endpoints.